+++ packages/connectivity_plus/CHANGELOG.md
@@ -1,3 +1,10 @@
 ## 1.0.0
 
-* Initial release
+* Initial release.
+
+## 1.1.0
+
+* Update connectivity_plus to 2.1.0.
+* Update connectivity_plus_platform_interface to 1.1.1.
+* Support ethernet as connectivity result.
+* Remove obsolete `connectivity_plus_tizen.dart`.